continued merge fixes
